SW-5807 Add NewApplicationModal to home page and open when the new application card CTA is clicked on
