+++ src/scenes/ApplicationRouter/ApplicationsList.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useState } from 'react';
+import React, { useState } from 'react';
 
 import { Box, Container, Grid } from '@mui/material';
 import { useDeviceInfo } from '@terraware/web-components/utils';
@@ -7,21 +7,15 @@ import { DateTime } from 'luxon';
 import ApplicationCard from 'src/components/Application/ApplicationCard';
 import PageHeader from 'src/components/PageHeader';
 import Button from 'src/components/common/button/Button';
-import useNavigateTo from 'src/hooks/useNavigateTo';
-import { useLocalization } from 'src/providers';
 import strings from 'src/strings';
-import useSnackbar from 'src/utils/useSnackbar';
 
 import { useApplicationData } from '../../providers/Application/Context';
 import NewApplicationModal from './NewApplicationModal';
 
 const ApplicationListView = () => {
-  const { activeLocale } = useLocalization();
   const { isTablet, isMobile } = useDeviceInfo();
-  const { toastSuccess } = useSnackbar();
   const { allApplications } = useApplicationData();
   const [isNewApplicationModalOpen, setIsNewApplicationModalOpen] = useState<boolean>(false);
-  const { goToApplication } = useNavigateTo();
 
   const primaryGridSize = () => {
     if (isMobile) {
@@ -40,16 +34,6 @@ const ApplicationListView = () => {
     return 6;
   };
 
-  const onApplicationCreated = useCallback(
-    (applicationId: number) => {
-      if (activeLocale) {
-        toastSuccess(strings.SUCCESS);
-      }
-      goToApplication(applicationId);
-    },
-    [activeLocale, goToApplication, toastSuccess]
-  );
-
   return (
     <Box
       component='main'
@@ -67,11 +51,7 @@ const ApplicationListView = () => {
           }
         />
         <Container maxWidth={false} sx={{ padding: 0 }}>
-          <NewApplicationModal
-            open={isNewApplicationModalOpen}
-            onClose={() => setIsNewApplicationModalOpen(false)}
-            onApplicationCreated={onApplicationCreated}
-          />
+          <NewApplicationModal open={isNewApplicationModalOpen} onClose={() => setIsNewApplicationModalOpen(false)} />
           <Grid container spacing={3} sx={{ padding: 0 }}>
             {allApplications?.map((application) => (
               <Grid
