+++ src/scenes/ApplicationRouter/NewApplicationModal.tsx
@@ -6,6 +6,7 @@ import { BusySpinner, Dropdown } from '@terraware/web-components';
 import DialogBox from 'src/components/common/DialogBox/DialogBox';
 import TextField from 'src/components/common/Textfield/Textfield';
 import Button from 'src/components/common/button/Button';
+import useNavigateTo from 'src/hooks/useNavigateTo';
 import { useProjects } from 'src/hooks/useProjects';
 import { useLocalization, useOrganization } from 'src/providers';
 import {
@@ -19,6 +20,7 @@ import {
 import { useAppDispatch, useAppSelector } from 'src/redux/store';
 import strings from 'src/strings';
 import useForm from 'src/utils/useForm';
+import useSnackbar from 'src/utils/useSnackbar';
 
 import { useApplicationData } from '../../providers/Application/Context';
 
@@ -31,17 +33,17 @@ type NewApplication = {
 export type NewApplicationModalProps = {
   open: boolean;
   onClose: () => void;
-  onApplicationCreated: (applicationId: number) => void;
 };
 
-const NewApplicationModal = ({ open, onClose, onApplicationCreated }: NewApplicationModalProps): JSX.Element => {
+const NewApplicationModal = ({ open, onClose }: NewApplicationModalProps): JSX.Element => {
   const theme = useTheme();
-
   const { activeLocale } = useLocalization();
   const { availableProjects } = useProjects();
   const { selectedOrganization } = useOrganization();
   const { allApplications } = useApplicationData();
   const dispatch = useAppDispatch();
+  const { toastSuccess } = useSnackbar();
+  const { goToApplication } = useNavigateTo();
 
   const [isLoading, setIsLoading] = useState<boolean>(false);
 
@@ -144,6 +146,16 @@ const NewApplicationModal = ({ open, onClose, onApplicationCreated }: NewApplica
     setCreateProjectApplicationRequestId,
     setIsLoading,
   ]);
+
+  const onApplicationCreated = useCallback(
+    (applicationId: number) => {
+      if (activeLocale) {
+        toastSuccess(strings.SUCCESS);
+      }
+      goToApplication(applicationId);
+    },
+    [activeLocale, goToApplication, toastSuccess]
+  );
 
   useEffect(() => {
     if (createApplicationResult && createApplicationResult.status === 'success' && createApplicationResult.data) {
