+++ src/scenes/Home/TerrawareHomeView.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useState } from 'react';
 
 import { Box, Container, Grid } from '@mui/material';
 import { useDeviceInfo } from '@terraware/web-components/utils';
@@ -11,10 +11,14 @@ import { useOrganization, useUser } from 'src/providers';
 import strings from 'src/strings';
 import { isAdmin } from 'src/utils/organization';
 
+import NewApplicationModal from '../ApplicationRouter/NewApplicationModal';
+
 const TerrawareHomeView = () => {
   const { user } = useUser();
   const { selectedOrganization } = useOrganization();
   const { isTablet, isMobile } = useDeviceInfo();
+
+  const [isNewApplicationModalOpen, setIsNewApplicationModalOpen] = useState<boolean>(false);
 
   const primaryGridSize = () => {
     if (isMobile) {
@@ -34,86 +38,91 @@ const TerrawareHomeView = () => {
   };
 
   return (
-    <Box
-      component='main'
-      sx={{
-        minHeight: '100vh',
-        display: 'flex',
-        flexDirection: 'column',
-      }}
-    >
-      <Box paddingRight={'24px'} paddingLeft={isMobile ? '24px' : 0}>
-        <PageHeader
-          title={user?.firstName ? strings.formatString(strings.WELCOME_PERSON, user.firstName) : strings.WELCOME}
-          subtitle=''
-        />
-        <Container maxWidth={false} sx={{ padding: 0 }}>
-          <Grid container spacing={3} sx={{ padding: 0 }}>
-            {isAdmin(selectedOrganization) && (
-              <>
-                <Grid item xs={primaryGridSize()}>
-                  <PageCard
-                    id='peopleHomeCard'
-                    name={strings.PEOPLE}
-                    icon='person'
-                    description={strings.PEOPLE_CARD_DESCRIPTION}
-                    link={APP_PATHS.PEOPLE}
-                    linkText={strings.formatString(strings.GO_TO, strings.PEOPLE) as string}
-                    linkStyle={'plain'}
-                  />
-                </Grid>
-                <Grid item xs={primaryGridSize()}>
-                  <PageCard
-                    id='seedbankHomeCard'
-                    name={strings.SEED_BANKS}
-                    icon='seedbankNav'
-                    description={strings.SEED_BANKS_CARD_DESCRIPTION}
-                    link={APP_PATHS.SEED_BANKS}
-                    linkText={strings.formatString(strings.GO_TO, strings.SEED_BANKS) as string}
-                    linkStyle={'plain'}
-                  />
-                </Grid>
-              </>
-            )}
-            <Grid item xs={secondaryGridSize()}>
-              <PageCard
-                id='speciesHomeCard'
-                name={strings.SPECIES}
-                icon='species'
-                description={strings.SPECIES_CARD_DESCRIPTION}
-                link={APP_PATHS.SPECIES}
-                linkText={strings.formatString(strings.GO_TO, strings.SPECIES) as string}
-                linkStyle={'plain'}
-              />
-            </Grid>
-            <Grid item xs={secondaryGridSize()}>
-              <PageCard
-                id='accessionsHomeCard'
-                name={strings.ACCESSIONS}
-                icon='seeds'
-                description={strings.ACCESSIONS_CARD_DESCRIPTION}
-                link={APP_PATHS.ACCESSIONS}
-                linkText={strings.formatString(strings.GO_TO, strings.ACCESSIONS) as string}
-                linkStyle={'plain'}
-              />
-            </Grid>
-            {isEnabled('Accelerator Application') && (
+    <>
+      <NewApplicationModal open={isNewApplicationModalOpen} onClose={() => setIsNewApplicationModalOpen(false)} />
+
+      <Box
+        component='main'
+        sx={{
+          minHeight: '100vh',
+          display: 'flex',
+          flexDirection: 'column',
+        }}
+      >
+        <Box paddingRight={'24px'} paddingLeft={isMobile ? '24px' : 0}>
+          <PageHeader
+            title={user?.firstName ? strings.formatString(strings.WELCOME_PERSON, user.firstName) : strings.WELCOME}
+            subtitle=''
+          />
+          <Container maxWidth={false} sx={{ padding: 0 }}>
+            <Grid container spacing={3} sx={{ padding: 0 }}>
+              {isAdmin(selectedOrganization) && (
+                <>
+                  <Grid item xs={primaryGridSize()}>
+                    <PageCard
+                      id='peopleHomeCard'
+                      name={strings.PEOPLE}
+                      icon='person'
+                      description={strings.PEOPLE_CARD_DESCRIPTION}
+                      link={APP_PATHS.PEOPLE}
+                      linkText={strings.formatString(strings.GO_TO, strings.PEOPLE) as string}
+                      linkStyle={'plain'}
+                    />
+                  </Grid>
+                  <Grid item xs={primaryGridSize()}>
+                    <PageCard
+                      id='seedbankHomeCard'
+                      name={strings.SEED_BANKS}
+                      icon='seedbankNav'
+                      description={strings.SEED_BANKS_CARD_DESCRIPTION}
+                      link={APP_PATHS.SEED_BANKS}
+                      linkText={strings.formatString(strings.GO_TO, strings.SEED_BANKS) as string}
+                      linkStyle={'plain'}
+                    />
+                  </Grid>
+                </>
+              )}
               <Grid item xs={secondaryGridSize()}>
                 <PageCard
-                  id='applicationHomeCard'
-                  name={strings.APPLY_TO_ACCELERATOR}
-                  icon='iconFile'
-                  description={strings.APPLY_TO_ACCELERATOR}
-                  link={APP_PATHS.APPLICATIONS}
-                  linkText={strings.START_NEW_APPLICATION}
-                  linkStyle={'button-primary'}
+                  id='speciesHomeCard'
+                  name={strings.SPECIES}
+                  icon='species'
+                  description={strings.SPECIES_CARD_DESCRIPTION}
+                  link={APP_PATHS.SPECIES}
+                  linkText={strings.formatString(strings.GO_TO, strings.SPECIES) as string}
+                  linkStyle={'plain'}
                 />
               </Grid>
-            )}
-          </Grid>
-        </Container>
+              <Grid item xs={secondaryGridSize()}>
+                <PageCard
+                  id='accessionsHomeCard'
+                  name={strings.ACCESSIONS}
+                  icon='seeds'
+                  description={strings.ACCESSIONS_CARD_DESCRIPTION}
+                  link={APP_PATHS.ACCESSIONS}
+                  linkText={strings.formatString(strings.GO_TO, strings.ACCESSIONS) as string}
+                  linkStyle={'plain'}
+                />
+              </Grid>
+              {isEnabled('Accelerator Application') && (
+                <Grid item xs={secondaryGridSize()}>
+                  <PageCard
+                    id='applicationHomeCard'
+                    name={strings.APPLY_TO_ACCELERATOR}
+                    icon='iconFile'
+                    description={strings.APPLY_TO_ACCELERATOR}
+                    link={APP_PATHS.APPLICATIONS}
+                    linkText={strings.START_NEW_APPLICATION}
+                    linkStyle={'button-primary'}
+                    onClick={() => setIsNewApplicationModalOpen(true)}
+                  />
+                </Grid>
+              )}
+            </Grid>
+          </Container>
+        </Box>
       </Box>
-    </Box>
+    </>
   );
 };
